--- internal/domain/user/model.go
@@ -1,9 +1,0 @@
-package user
-
-import "github.com/google/uuid"
-
-type User struct {
-	ID       uuid.UUID `json:"id"`
-	Name     string    `json:"userName"`
-	Password string    `json:"password"`
-}
