+++ internal/server/routes.go
@@ -4,17 +4,17 @@ import (
 	"net/http"
 
 	"github.com/go-chi/chi"
-	"github.com/tullur/lets-go-chat/internal/domain/user"
+	"github.com/tullur/lets-go-chat/internal/domain/user/memory"
 	"github.com/tullur/lets-go-chat/internal/handlers"
 )
 
 func UserRoutes() http.Handler {
 	r := chi.NewRouter()
-	userStore := user.New()
+	userInMemoryStore := memory.NewInMemoryRepository()
 
-	r.Get("/", handlers.HandleUserList(userStore))
-	r.Post("/", handlers.HandleUserCreation(userStore))
-	r.Post("/login", handlers.HandleUserLogin(userStore))
+	r.Get("/", handlers.HandleUserList(userInMemoryStore))
+	r.Post("/", handlers.HandleUserCreation(userInMemoryStore))
+	r.Post("/login", handlers.HandleUserLogin(userInMemoryStore))
 
 	return r
 }
